Split class PlatformThreads off from JavaThreads.
